0x14-bit_manipulation

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -14,7 +14,7 @@ int get_bit(unsigned long int n, unsigned int index)
 {
 	unsigned long int mask = 1UL << index;
 
-	if (index >= sizeof(unsigned int) * 8)
+	if (index >= sizeof(unsigned long int) * 8)
 		return (0);
 	return ((n & mask) != 0);
 }
